% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luaScript.R
\name{luaScript}
\alias{luaScript}
\title{Add a Lua script to a Shiny UI or R Markdown Document}
\usage{
luaScript(lua = NULL, src = NULL)
}
\arguments{
\item{lua}{Character string containing lua code. Do not set a value for this argument if using Lua from a separate source file.}

\item{src}{Relative path to a Lua source file in the Shiny app's 'www' directory.}
}
\description{
Generates the html to embed a client-side Lua script in a Shiny UI or an R Markdown
webpage. Lua code can be passed in with a string or with a separate Lua file in the app's www directory
}
\details{
\code{\link{useLuaVM}} must be run in the Shiny app's UI for this function to work.

To add a Lua script to a shiny UI without this function, use \code{tags$script()} with
\code{type = "application/lua"}
}
