% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useLuaVM.R
\name{useLuaVM}
\alias{useLuaVM}
\title{Add Lua scripting capability to the UI of a Shiny app}
\usage{
useLuaVM(rmd = FALSE)
}
\arguments{
\item{rmd}{Logical; this must be set to \code{TRUE} if you are using \code{shinyLua}
in an interactive R Markdown document rather than a Shiny app.}
}
\description{
Add Lua scripting capability to the UI of a Shiny app
}
