% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useLuaVM.R
\name{useLuaVM}
\alias{useLuaVM}
\title{Add Lua scripting capability to the UI of a Shiny app}
\usage{
useLuaVM(rmd = FALSE)
}
\arguments{
\item{rmd}{Logical; this must be set to \code{TRUE} if you are using \code{shinyLua}
in an interactive R Markdown document rather than a Shiny app.}
}
\description{
This is a setup function that must be called from a Shiny app's UI to
add the Lua implementation required for other \code{shinyLua} functions to work.
}
\details{
This function adds 'fengari-web.js', a JavaScript port of the Lua VM, to an app's <head> tag
and resource path along with other startup scripts and dependencies. Any scripts in the UI
of type \code{"application/lua"} will run in this embedded Lua VM. \code{shinyLua} provides a
convenience function, \code{\link{luaScript}}, for inserting Lua scripts into a Shiny UI.
}
